backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic>=2.10.1,<3.0.0
pydantic-settings>=2.6.1
python-dotenv==1.0.1
httpx==0.27.2

SQLAlchemy==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2

mcp[cli]==1.2.0
google-generativeai==0.8.3

# Optional integrations
slack_sdk==3.27.1
google-api-python-client==2.129.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1

email-validator==2.2.0 